+++ Resources/js/React/PageTemplates/11.js
@@ -1,10 +1,10 @@
-﻿import React, { Component, Fragment } from 'react';
+﻿import React, { Component } from 'react';
+import GoogleMap from '../CMS-Blocks/GoogleMap';
 import PageTitle from '../CMS-Blocks/PageTitle';
 import TextBlock from '../CMS-Blocks/Text';
-import MediaBlock from '../CMS-Blocks/Media';
+import EmailForm from '../CMS-Blocks/EmailForm';
 
-
-export default class ReactTemplate11 extends Component {
+export default class ReactTemplate3 extends Component {
     constructor(props) {
         super(props);
 
@@ -17,19 +17,16 @@ export default class ReactTemplate11 extends Component {
 
     render() {
         return (
-            <Fragment>
-                <div className="row">
-                    <div className="col-12">
-                        <PageTitle title={this.props.data.name}
-                            public={this.props.data.public}
-                            created={this.props.data.created}
-                            displayAdmin={!!this.props.data.enums}
-                        />
-                    </div>
-                </div>
+            <React.Fragment>
+                <PageTitle title={this.props.data.name}
+                    public={this.props.data.public}
+                    created={this.props.data.created}
+                    displayAdmin={!!this.props.data.enums}
+                />
 
-                <div className="row mb-5">
-                    <div className="col-md-6 col-sm-12">
+
+                <div className="row pb-4">
+                    <div className="col-lg-4 col-md-6 col-sm-12">
                         <TextBlock allowEdits={this.props.allowEdits}
                             content={this.props.data.textComponents[0] || null}
                             pushChanges={this.props.pushChanges.bind(this, 'text')}
@@ -37,33 +34,15 @@ export default class ReactTemplate11 extends Component {
                         />
                     </div>
 
-                    <div className="col-md-6 col-sm-12">
+                    <div className="col-lg-4 col-md-6 col-sm-12">
                         <TextBlock allowEdits={this.props.allowEdits}
                             content={this.props.data.textComponents[1] || null}
                             pushChanges={this.props.pushChanges.bind(this, 'text')}
                             settings={this.props.data.enums}
                         />
                     </div>
-                </div>
 
-                <div className="row mb-5">
-                    <div className="col-md-6 col-sm-12">
-                        <MediaBlock allowEdits={this.props.allowEdits}
-                            content={this.props.data.mediaComponents[0] || null}
-                            pushChanges={this.props.pushChanges.bind(this, 'media')}
-                        />
-                    </div>
-
-                    <div className="col-md-6 col-sm-12">
-                        <MediaBlock allowEdits={this.props.allowEdits}
-                            content={this.props.data.mediaComponents[1] || null}
-                            pushChanges={this.props.pushChanges.bind(this, 'media')}
-                        />
-                    </div>
-                </div>
-
-                <div className="row mb-5">
-                    <div className="col-12">
+                    <div className="col-lg-4 col-md-6 col-sm-12">
                         <TextBlock allowEdits={this.props.allowEdits}
                             content={this.props.data.textComponents[2] || null}
                             pushChanges={this.props.pushChanges.bind(this, 'text')}
@@ -71,7 +50,18 @@ export default class ReactTemplate11 extends Component {
                         />
                     </div>
                 </div>
-            </Fragment>
+
+                <div className="row pt-2">
+                    <div className="col-lg-6 col-md-12 pb-2">
+                        <EmailForm config={this.props.data.otherComponents.captchaSiteKey}/>
+                    </div>
+
+                    <div className="col-lg-6 col-md-12">
+                        <GoogleMap title="Where to Find Us"
+                            config={this.props.data.otherComponents.googleMaps || null } />
+                    </div>
+                </div>
+            </React.Fragment>
         );
     }
 }
