+++ Resources/js/React/PageTemplates/2.js
@@ -1,11 +1,10 @@
-﻿import React, { Component } from 'react';
+﻿import React, { Component, Fragment } from 'react';
 import PageTitle from '../CMS-Blocks/PageTitle';
 import TextBlock from '../CMS-Blocks/Text';
-import Media from '../CMS-Blocks/Media';
+import MediaBlock from '../CMS-Blocks/Media';
 
-const baseUri = "/api/page"
 
-export default class ReactTemplate2 extends Component {
+export default class ReactTemplate4 extends Component {
     constructor(props) {
         super(props);
 
@@ -15,12 +14,13 @@ export default class ReactTemplate2 extends Component {
         }
     }
 
+
     render() {
         return (
-            <React.Fragment>
+            <Fragment>
                 <div className="row">
                     <div className="col-12 px-0 mb-4" style={{ 'marginTop': '-1rem' }}>
-                        <Media allowEdits={this.props.allowEdits}
+                        <MediaBlock allowEdits={this.props.allowEdits}
                             content={this.props.data.mediaComponents[0] || null}
                             pushChanges={this.props.pushChanges.bind(this, 'media')}
                         />
@@ -34,7 +34,7 @@ export default class ReactTemplate2 extends Component {
                         />
                     </div>
 
-                    <div className="col-lg-6 col-md-12 py-2">
+                    <div className="col-12 py-2">
                         <TextBlock allowEdits={this.props.allowEdits}
                             content={this.props.data.textComponents[0] || null}
                             pushChanges={this.props.pushChanges.bind(this, 'text')}
@@ -42,7 +42,21 @@ export default class ReactTemplate2 extends Component {
                         />
                     </div>
 
-                    <div className="col-lg-6 col-md-12">
+                    <div className="col-md-6 col-sm-12 mb-2">
+                        <MediaBlock allowEdits={this.props.allowEdits}
+                            content={this.props.data.mediaComponents[1] || null}
+                            pushChanges={this.props.pushChanges.bind(this, 'media')}
+                        />
+                    </div>
+
+                    <div className="col-md-6 col-sm-12 mb-2">
+                        <MediaBlock allowEdits={this.props.allowEdits}
+                            content={this.props.data.mediaComponents[2] || null}
+                            pushChanges={this.props.pushChanges.bind(this, 'media')}
+                        />
+                    </div>
+
+                    <div className="col-12 py-2">
                         <TextBlock allowEdits={this.props.allowEdits}
                             content={this.props.data.textComponents[1] || null}
                             pushChanges={this.props.pushChanges.bind(this, 'text')}
@@ -50,7 +64,7 @@ export default class ReactTemplate2 extends Component {
                         />
                     </div>
                 </div>
-            </React.Fragment>
+            </Fragment>
         );
     }
 }
