+++ Resources/js/React/PageTemplates/5.js
@@ -4,7 +4,7 @@ import TextBlock from '../CMS-Blocks/Text';
 import MediaBlock from '../CMS-Blocks/Media';
 
 
-export default class ReactTemplate5 extends Component {
+export default class ReactTemplate10 extends Component {
     constructor(props) {
         super(props);
 
@@ -19,13 +19,6 @@ export default class ReactTemplate5 extends Component {
         return (
             <Fragment>
                 <div className="row">
-                    <div className="col-12 px-0 mb-4" style={{ 'marginTop': '-1rem' }}>
-                        <MediaBlock allowEdits={this.props.allowEdits}
-                            content={this.props.data.mediaComponents[0] || null}
-                            pushChanges={this.props.pushChanges.bind(this, 'media')}
-                        />
-                    </div>
-
                     <div className="col-12">
                         <PageTitle title={this.props.data.name}
                             public={this.props.data.public}
@@ -33,8 +26,10 @@ export default class ReactTemplate5 extends Component {
                             displayAdmin={!!this.props.data.enums}
                         />
                     </div>
+                </div>
 
-                    <div className="col-md-6 col-sm-12 mb-4">
+                <div className="row mb-5">
+                    <div className="col-md-6 col-sm-12">
                         <TextBlock allowEdits={this.props.allowEdits}
                             content={this.props.data.textComponents[0] || null}
                             pushChanges={this.props.pushChanges.bind(this, 'text')}
@@ -42,14 +37,16 @@ export default class ReactTemplate5 extends Component {
                         />
                     </div>
 
-                    <div className="col-md-6 col-sm-12 mb-4">
+                    <div className="col-md-6 col-sm-12">
                         <MediaBlock allowEdits={this.props.allowEdits}
-                            content={this.props.data.mediaComponents[1] || null}
+                            content={this.props.data.mediaComponents[0] || null}
                             pushChanges={this.props.pushChanges.bind(this, 'media')}
                         />
                     </div>
+                </div>
 
-                    <div className="col-md-6 col-sm-12 mb-4">
+                <div className="row mb-5">
+                    <div className="col-md-6 col-sm-12">
                         <TextBlock allowEdits={this.props.allowEdits}
                             content={this.props.data.textComponents[1] || null}
                             pushChanges={this.props.pushChanges.bind(this, 'text')}
@@ -57,7 +54,24 @@ export default class ReactTemplate5 extends Component {
                         />
                     </div>
 
-                    <div className="col-md-6 col-sm-12 mb-4">
+                    <div className="col-md-6 col-sm-12">
+                        <MediaBlock allowEdits={this.props.allowEdits}
+                            content={this.props.data.mediaComponents[1] || null}
+                            pushChanges={this.props.pushChanges.bind(this, 'media')}
+                        />
+                    </div>
+                </div>
+
+                <div className="row mb-5">
+                    <div className="col-md-6 col-sm-12">
+                        <TextBlock allowEdits={this.props.allowEdits}
+                            content={this.props.data.textComponents[2] || null}
+                            pushChanges={this.props.pushChanges.bind(this, 'text')}
+                            settings={this.props.data.enums}
+                        />
+                    </div>
+
+                    <div className="col-md-6 col-sm-12">
                         <MediaBlock allowEdits={this.props.allowEdits}
                             content={this.props.data.mediaComponents[2] || null}
                             pushChanges={this.props.pushChanges.bind(this, 'media')}
