+++ Resources/js/React/PageTemplates/6.js
@@ -4,7 +4,7 @@ import TextBlock from '../CMS-Blocks/Text';
 import MediaBlock from '../CMS-Blocks/Media';
 
 
-export default class ReactTemplate6 extends Component {
+export default class ReactTemplate11 extends Component {
     constructor(props) {
         super(props);
 
@@ -26,8 +26,10 @@ export default class ReactTemplate6 extends Component {
                             displayAdmin={!!this.props.data.enums}
                         />
                     </div>
+                </div>
 
-                    <div className="col-md-6 col-sm-12 mb-4">
+                <div className="row mb-5">
+                    <div className="col-md-6 col-sm-12">
                         <TextBlock allowEdits={this.props.allowEdits}
                             content={this.props.data.textComponents[0] || null}
                             pushChanges={this.props.pushChanges.bind(this, 'text')}
@@ -35,16 +37,35 @@ export default class ReactTemplate6 extends Component {
                         />
                     </div>
 
-                    <div className="col-md-6 col-sm-12 mb-4">
+                    <div className="col-md-6 col-sm-12">
+                        <TextBlock allowEdits={this.props.allowEdits}
+                            content={this.props.data.textComponents[1] || null}
+                            pushChanges={this.props.pushChanges.bind(this, 'text')}
+                            settings={this.props.data.enums}
+                        />
+                    </div>
+                </div>
+
+                <div className="row mb-5">
+                    <div className="col-md-6 col-sm-12">
+                        <MediaBlock allowEdits={this.props.allowEdits}
+                            content={this.props.data.mediaComponents[0] || null}
+                            pushChanges={this.props.pushChanges.bind(this, 'media')}
+                        />
+                    </div>
+
+                    <div className="col-md-6 col-sm-12">
                         <MediaBlock allowEdits={this.props.allowEdits}
                             content={this.props.data.mediaComponents[1] || null}
                             pushChanges={this.props.pushChanges.bind(this, 'media')}
                         />
                     </div>
+                </div>
 
+                <div className="row mb-5">
                     <div className="col-12">
                         <TextBlock allowEdits={this.props.allowEdits}
-                            content={this.props.data.textComponents[1] || null}
+                            content={this.props.data.textComponents[2] || null}
                             pushChanges={this.props.pushChanges.bind(this, 'text')}
                             settings={this.props.data.enums}
                         />
